[MT-1201] QueueManager (#263)
* [WIP] Create SQLQueueRepository

Update TealiumDispatch to store [String: Any] to be created from a local json
Always use Int64 instead of Double for timestamps
Use Tealium.jsonEncoder and Tealium.jsonDecoder

Fix Expiry to always use milliseconds

* Add queue repository tests

* Add more tests and add a basic implementation to limit the enqueue size to the maxQueueSize

* Add timestamp and expiry tests

* Experiment with consent queue in a different table

* Make consent use the same queues as the dispatchers

Make all dispatchers in the DB be referred as processors instead to be more generic

* Implement QueueRepository in QueueManager

and use QueueManager for DispatchManager and ConsentManager

* Switch queue manager to use TealiumMutableState for inflightEvents

* Add dispatchers list in consent manager

* Fix consent manager logic according to new queuemanager

* Make ConsentManager a module in order to be used with QueueManager processors list

This also helps with reusing settings

Updated also Barrier and Transformer Coordinators to support runtime additional barriers/transformers to be added by modules that are enabled/disabled

* Add QueueManager tests and make QueueManager check for ConsentManager instead of ConsentModule for testing

* Add tests for multiple Dispatches and prevent duplicate tracking of the same event

Add some documentation comments

* Add logs for QueueManager and move processors filtering outside

* Refactor TealiumMutableState and TealiumObservableState to be actual Subject/Observable and rename them into TealiumVariableSubject and TealiumStatefulObservable

This was done to be sure to treat them in the same way as the others and only report subscription up the chain.
In the old system you could subscribe to an observable state and not receive updates due to the state itself being deinitialized.

* Update QueueManager logs

* Add failure QueueManager tests

* Make dispatches insert or replace

Improve methods signature

* Update QueueManager logs

* Rename Events to Dispatches in QueueManager

* Rename deleteDispatches(nonQueuedFor) in deleteQueues(forProcessorsNotIn)

Update documentation

* Update deleteQueue method signature and logs in QueueManager

* Fix QueueManager receiving an empty processors list at the launch of the application and remove extra delete called for empty dispatches

* Fix TealiumImplementation tests and add a test for empty modules

* Update QueueManager logs

* Use BarrierRegistry and TransformerRegistry in the TealiumContext to limit coordinator usage in modules

* Fix QueueManager expiration filter

* Remove useless projectedValue from TealiumVariableState

Make sure modules proessors are distinct when sent to the QueueManager

* Make queue manager react to core settings changes and delete expired items on launch and settings change

* Improve QueueManager test names and ConsentModule logic

* Update log message in QueueManager

+++ tealium-swift/core/Persistence/DataModels/ExpirationRequest.swift
@@ -12,7 +12,7 @@ enum ExpirationRequest {
     case restart
     case sessionChange
 
-    var expiryTime: Double {
+    var expiryTime: Int64 {
         switch self {
         case .restart:
             return Expiry.untilRestart.expiryTime()
